backend/user-service/requirements.txt
Django>=5.2.3,<5.3
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-decouple==3.8
djangorestframework-simplejwt==5.5.1